#6 split the code in different files.
